requirements.txt
# Atlas Fabric Requirements
# Core dependencies for the benchmarking framework

# Required dependencies
PyYAML>=6.0.1

# Optional dependencies for visualization
# Uncomment the following lines to enable visualization features:
# matplotlib>=3.7.0
# numpy>=1.24.0

# Optional dependencies for web interface (future feature)
# streamlit>=1.28.0
# plotly>=5.17.0

# Optional dependencies for data export (future feature)
# pandas>=2.0.0
# openpyxl>=3.1.0
# xlsxwriter>=3.1.0

# Development dependencies
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
# pre-commit>=3.4.0
